Reworked the video controls interface to make it cleaner
